+++ django_typer/__init__.py
@@ -47,7 +47,7 @@ from .management import (
     model_parser_completer,  # noqa: F401
 )
 
-VERSION = (2, 5, 0)
+VERSION = (2, 6, 0)
 
 __title__ = "Django Typer"
 __version__ = ".".join(str(i) for i in VERSION)